requirements.txt
icecream
python-vagrant
